zzhang

+++ dltCalibration.cpp
@@ -18,13 +18,13 @@ using namespace cv;
 
 int main(int argc, const char* argv[]) 
 {
-    std::vector<Point2f> worldpts = prasePointSFromFile("data\\corners_real.dat");
-    std::vector<Point2f> imagepts = prasePointSFromFile("data\\corners_1.dat");
+    std::vector<Point2d> worldpts = prasePointSFromFile("data\\corners_real.dat");
+    std::vector<Point2d> imagepts = prasePointSFromFile("data\\corners_1.dat");
 
     Mat homoMat =    findHomography(worldpts, imagepts);
     cout << "homoMat:" << homoMat << endl;
 
-    float homoMat2[9];
+    double homoMat2[9];
     computeHomography(worldpts, imagepts, homoMat2);
     for (int i =0; i < 9;i++)
     {
