+++ dltalgo.h
@@ -1,3 +1,3 @@
 #pragma once
 
-bool computeHomography(std::vector<Point2f>& srcPts, std::vector<Point2f> dstPts, float homograpy[9]);
+bool computeHomography(std::vector<Point2d>& srcPts, std::vector<Point2d> dstPts, double homograpy[9]);
